CA-288635: increase db flush chunk size
The semantics we actually want when flushing the entire DB is to
write out all the chunks and then wait until the operation is finished.
We can't do that since the file is opened with O_DSYNC, due to the lack
of an `fdatasync` in the Unix module.

To reduce the latency penalty from waiting for each 16KiB chunk to be
written increase the chunk size. Internally the ocaml runtime splits
writes greater than 64KiB into 64KiB chunks, so use that as our
blocksize.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>