CA-289140: log the exception from PBD plug without doing other API calls
When Xapi_pbd.plug fails we called an API to retrieve the UUID which can
fail and then we loose the original exception before we log it.
We already have the UUID in the PBD record, so log that.

In this case both errors were likely the same (the PBD did not exist
anymore, hence looking it up would fail and plug would fail) though.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>