CA-289140: ignore missing PBDs on startup
If SR.forget is run after we got a list of PBDs, in plug_all_pbds then
there is a race condition where by the time we try to plug the PBD it
won't exist anymore.
If you have a lot of PBDs, or the PBD plugs are slow this race condition
can be hit, and it prevents the host from getting enabled after boot.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>