CP-27172: integrate Host.destroy and Host.declare_dead with clustering
Declare dead requires the rest of the hosts in the cluster to be alive
(otherwise data corruption could occur as we are about to forcibly kick
 hosts from the cluster).
The current Xen API provides a way to kick hosts out of a pool one by
one, or to mark them as dead. We could introduce a batched API, but that
would be cumbersome for the user, it is better if we can declare dead
hosts one at a time, and when we declared them all dead then we actually
modify the cluster.

The semantics of "declare dead" is different between XenAPI and
clustering daemon, for XenAPI it means: the host is dead now, you can
reset VM power state, etc. In fact HA automatically detects dead hosts.

Then there is Host.destroy (or `host-forget`) on the CLI which considers
the host permanently dead. This is the semantics the clustering daemon
wants: if a host that has been declared dead to the clustering daemon
comes back alive then data corruption could occur.

To avoid confusion, call the new method `Cluster_host.forget`, and call
it automatically from `Host.destroy`.
`Host.declare_dead` will have no effect on clustering.

For this to work we need to remember all the IPs of the hosts we have
attempted to declare dead in the past but failed, and retry declaring
them as dead again on the next API call.

Declaring a host as dead is permanent, the host can never be part of the
cluster again, so we do not need an undo API here.

TODO: although things could go wrong if we try to join a host with the same IP
as one that we are trying to declare as dead... (which can happen due to DHCP).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>