BP CP-28117: introduce helper module for running tasks in parallel
HA VM restarts need to happen in the order defined by the HA policy and
handle failures.  The initial implementation used code specialized to
do that by preparing a record with VMs and tasks, sorting them, retrying
them, but it soon became unreadable and each slight modification required
large restructuring in the code.

The code is simpler and more readable if we introduce a general purpose
mechanism for performing actions after a task has finished, and for
automatically grouping actions by an order number and launching them
in parallel. The abstraction used is a monad, however the knowledge of
monads is kept minimal by having the naming and documentation in the
code refer to tasks and actions, rather than abstract monad concepts.
The code tries to be careful about exceptions, and wraps them in a
Result type.

This is the just the helper module implementation.

Note for backporting:
 older versions will not have a result module, you need to define:
 ```
 module Result = struct type 'a result = Ok of 'a | Error of 'a end
 let (|>) x f = f x
 ```

 `Xapi_stdext_std` is just `Stdext`
 'let open ... in' probably needs to be replaced with an open at toplevel

To make backporting easier GADTs are not used here, and changes are contained in a single file.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>