CP-24692: Prevent race condition between allowing PIF.unplug and clustering
`PIF.disallow_unplug:true` is necessary for the entire time a cluster
host is enabled, as well as being a prerequisite for the API calls
`Cluster_host.{enable, create}`.

If `disallow_unplug` is set to false during either of these two
clustering operations, it could create a race condition where both
operations are allowed, but we then have a cluster_host that doesn't
meet networking requirements and will therefore fail later on.

This PR introduces a condition wherein you cannot allow PIF.unplug if
the pool cluster is currently creating or enable a cluster_host on *any*
PIF, regardless of which PIF you are allowing unplug for. This should
rarely cause an issue, as `PIF.set_disallow_unplug` is rarely called,
although an operation like `Cluster.pool_{create, resync}` could block
it for a while, depending on the pool size.

Additionally, this commit removes the obsolete
`CLUSTERING_ENABLED_ON_NETWORK` API error in favour of using
`CLUSTERING_ENABLED`, as cluster_hosts correspond more closely to PIFs

Signed-off-by: Akanksha Mathur <akanksha.mathur@citrix.com>