CA-290471: fix wrong token timeout in XenCenter
Use same units in the object and in the API.
The timeout set by XenCenter 20000s, which is wrong, it should've been
20s (which is what you get from the CLI).
XenCenter creates an object and then reads the field to find out the
default, because the SDK doesn't expose the defaults for parameters.

So we should really use the same units for API parameters and internal
object fields. Especially that this field is not used by XAPI at all,
it is only there so that the admin can inspect what values were used
to create the cluster.

Changing this field will not change the actual timeouts used by the
cluster, because the timeout values cannot be changed at runtime.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>