CA-290006 update snapshot device model profile
A snapshot stores all relevant VM meta data inside VM.snapshot_metadata
which contains a platform value, which in turn contains a device model
profile. This needs to be updated when xapi's database is updated but so
far was not. Confusingly, the VM.platform field of a snapshot is not
used when restoring a snapshot. This commit updates the device model
profile inside the snapshot metadata.

Snapshot meta data is encoded as an s-expression. Updating it thus
requires decoding, updating, and encoding it again.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>