CP-28365: do not lose the exception text in log_and_ignore_exn
`log_and_ignore_exn` just called `log_backtrace` which gave us a backtrace
in the logs, but the actual exception message was lost, which can be
quite important for debugging.

Use the outer `log_backtrace_exn` which logs both the backtrace and the
message, using the improved backtraces saved by the backtrace module.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>