CP-28406: refactor Cluster/Cluster_host.destroy and forbid RPCs if clustering daemon is down
Cluster.destroy did not work if you destroyed the last cluster host
(like some of the unit tests actually did).
Cluster_host.destroy on the last node is special: there is no cluster
after you leave (leave would fail in reality), so just destroy the cluster.
Refactor all 3 destroy operations into one, choose automatically based
on number of hosts.

Could've introduced a new API error to forbid destroying the last
cluster host, but it is better if XAPI is able to automatically do the
right thing than to tell the user it should call some other API instead.

Also with Cluster_host.force_destroy you could have already ended up in
a situation where you have no cluster hosts and want to destroy the
cluster, which would hang indefinitely because the daemon was stopped.

We always try to enable the daemon on startup, so keep track on whether
we think it should be running, and if we know we stopped it then just
raise an error when trying to do an RPC.

This is useful in debugging situations where we try to send RPCs too
early too (e.g. before we started the daemon).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>