Rename Xapi_vm_snapshot.default_values to overrides
This better reflects what is it for. It is an assoc list of key/value pairs
that is passed to `copy_vm_fields`. When this function find a
`snapshot_metadata` key that is also in the `overrides` list, it will use the
value from `overrides` rather than the one from the snapshot.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>