CA-290874: Always derive domain_type from HVM_boot_policy on VM.revert
This addresses two related issues, depending on when the snapshot was taken.

Firstly, if the snapshot may have been taken on the Jura release, where the domain_type field was introduce, but always set to "unspecified". The snapshot_metadata will then include "domain_type=unspecified", and the resulting VM will do too.

If the VM was snapshotted on a release that predates the domain_type field,
then the snapshot_metadata will not include it. On revert, the VM will keep its
current domain_type, which may have changed since the snapshot was taken.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>