CA-287838: Slave Dom0 becomes incorrect state in DB after pool join
From the fix of CA-280981, we import the dom0 record of the slave into
the master rather than re-generating a new VM record for it during pool-join.
However it will make the dom0 record of slave into an incorrect state.

This fix will correct the dom0 record of the slave when xapi reatrting by:
1. set the power state to Running
2. set the domid to 0

Signed-off-by: Yang Qian <yang.qian@citrix.com>