CA-291017: Unable to connect server in pool of 64 physical hosts
Stunnel has a formula `max_clients=max_fds>=256 ? max_fds*125/256
: (max_fds-6)/2` to calculate how many clients can connect to this stunnel
server process, (https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=599138).
The default max_fds (max number of opened file descriptors) is 1024. That means
stunnel could only use 1024 * (125/256) = 500 file descriptors for client
connections. It is not enough for the master in the pool of 64 hosts when the pool
experiencing high loads.

Changes in this patch:
1. As xapi is restricted to use at most 1024 file descriptors, so increase
stunnel's process limit for the number of opened file descriptors to 2048 (so
it could use 2048 * (125/256) = 1000 file descriptors for client connections).
1.1 1000 stunnel client connections to the master is enough for the 8
connections * 63 slaves = 504 connections necessary in a pool of 64 hosts.
1.2 The remaining 496 possible stunnel client connections in the pool of 64 hosts
are still more than 380 previously when the ulimit=1024 (and therefore the connections
were limited to 500 - 8*15 = 380 in the pool of 16 hosts), so the system is not worse
off to take up extra xencenter,xe-cli and other stunnel client connections.
2. This change only affects the stunnel process on the server role
3. Stunnel uses `poll` rather than `select`, so this ulimit increase to 2048 is
fine

Signed-off-by: Yang Qian <yang.qian@citrix.com>