Generalize quicktest filtering
Now all the parameters of the test cases are provided by the filter.
This makes it possible to later add test that use VM templates, if the
required template is available.
Or to test migration between two SRs, both of which are automatically
selected by the filtering framework. This new framework also makes it
possible to generate every possible combination of SRs, and take up to
10 at random, and then test VDI migration / copy between them.

Now even passing the connection parameters (rpc & session_id) is handled
by this framework, using the "conn" filter.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>