CA-291197: Toolstack no longer ignores junk at the end of JSON-RPC response
The previous changes for CA-236855 no longer tolerant junk at the end of
input from the pvsproxy, which makes the interface more fragile. To build
a more robust interface in case of issue on the pvsproxy side, new option
strict from jsonrpc is used to interpret the response, so tailing junk
is ignored.

Signed-off-by: YarsinCitrix <yarsin.he@citrix.com>