CA-292372: Create SM objects for all running SMAPIv2 drivers
Previously we only created SM objects for those SMAPIv2 drivers that had
an SR of that type. This is inconsistent with the SMAPIv1 behaviour,
because we always create SMs for all the SMAPIv1 drivers.

This commit ensures that we create SMs for all the running SMAPIv2
drivers, even the ones we did not know about.

I've also made this function safer, because if it raises an exception,
it will cause the startup sequence to fail. Now the function will not
fail if we fail to query one of the SMAPIv2 plugins, it will just log an
error.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>