CA-292621: ensure PIF has an IP before calling clustering methods
The IP address would be updated asynchronously by the network monitor thread,
so we can detect there is an IP on the link but fail later when trying to read
the IP from the DB.
We cannot simply wait on the condition variable until the DB is updated:
the condition is not signalled. We could use the event mechanism to wait
until the IP field gets updated, but that adds more connections to the
master which hurts large pool scalability.
Also when we wake up we couldn't been woken up due to a change on
another PIF, not the one we were watching.

To avoid race conditions always update the IP addresses of all PIFs on a
networking change, and check for the presence of an IP address in the DB
in the clustering network watcher thread.
That way we know that when we get to the clustering code we will have an
IP address.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>