CA-292873: Allow "xl destroy" to work with QEMU upstream VMs
Xen documents /local/domain/%d/image/device-model-pid as the xenstore
path for QEMU's pid and libxl relies on this path to correctly destroy a
domain. Set this path in qemu-wrapper which makes "xl destroy" work with
QEMU upstream VMs.  qemu-trad watches for @releaseDomain itself which is
why "xl destroy" works with qemu-trad VMs.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>