CA-287333: Report storage backtraces in JSON, not S-expressions
Currently, attaching a VBD to a VDI on a GFS2 SR causing a storage
backend error, with an accompanying backtrace.

Unfortunately, while Xapi-storage-script reports this backtraces in JSON
format, Xenopsd expects an S-expression. This causes an S-expression parsing
failure, obfuscating the actual error and making debugging very difficult.

This commit patches Xenopsd's backtrace handling to expect a backtrace in JSON,
and should enable easier debugging of the VDI attach error.

Signed-off-by: Akanksha Mathur <akanksha.mathur@citrix.com>