CA-292641: Collect log messages from libs using Logs
We register our log reporter when an xcp service is configured, this way
we can report logs coming from libraries that use Logs. For example, we
can report errors in the cleanup function from stdext's finally.

For efficiency, we only consider logs of level warning and above from
libraries, to avoid calling our output_log function too often, which
does the actual filtering. It will be able to filter logs from libraries
too, the same way it filters them for toolstack components, because we
convert the Logs level and source to our own log level type and "brand".

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>