CA-287525 replace OPasswd with implementation in C
This replaces the OPasswd library for updating password entries with a
new implementation written in C that uses the libc API directly. The
main motivation is that OPasswd sometimes writes garbage to password
entries and we were not able to track this down. OPasswd uses CTypes to
bind C API functions. This commit implements the main functions in C and
exposes simple functions that are exported to OCaml using hand-written
glue code. We hope that this will be easier to debug in case of errors.

The code in unixpwd/ is ready to be released as an independent library.
However, it is easier to mature it inside the xen-api.git repository and
to do this at a later time.

The code comes with unit tests which can be run from unixpwd/Makefile.
For debugging, the C code also contains a minimal command line
application that can be built from c/Makefile.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>