CA-292063: detach static VDIs when HA is disarmed on boot
HA is disarmed on boot if we determine that it got disabled on the pool
while we were off.
However the static VDIs stayed attached, which prevented the PBD from
getting unplugged, e.g. on shutdown. Normally the shutdown code would unplug
these but since HA was disabled that code does not run.

Clean up any static VDIs at the point where we disarm HA on boot by
releasing all HA resources, not just the monitor.
The monitor is stopped when releasing the HA resources, so replace the
call to stop a monitor with releasing the resources.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>