CA-289650: Wait for the pidfile from udhcpd before releasing lock
The pidfile is written _after_ udhcpd daemonizes, so we need to wait
for a short amount of time until it turns up.

Extended explanation due to @edwintorok:

"the reason it was broken before was that xapi was looking at the pidfile to find old
udhcpds and kill them, however udhcpd takes a while to start up and write out a pidfile,
so if we try to run this function multiple times in quick succession we end up spawning
more udhcpd without realising. By waiting for the pidfile to be available we guarantee
that if we get called again a pidfile will be found and we can kill the process
properly."

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>