Only run SR.create compat fallback for PVS version of SM scripts
This makes the code stricter - previously we ran the compat code for all
versions, and as a result we accepted an empty response, even from newer
versions.

This change will make the code stricter, and probably safer, now we'll
fail if the newer storage scripts return an invalid, empty response from
SR.create.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>