CA-293858: Further prevention of logrotate running on old partition scheme
There's a firstboot script that attempts to prevent the new logrotate script
running when you've got the old-style partition layout. It does this by moving
the cron file aside. Unfortunately the cron file belongs to the xapi RPM and
hence installing any newer xapi-core RPM puts the file back in place.

This is a quick-fix solution that applies the same logic that the firstboot
script uses to detect old-style partitioning and bails out if it sees it.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>