CA-293399: Corrupted PCI information in xapi database
The language binding to the C function is:
foreign "pci_lookup_name"
    (Pci_access.t @-> ptr char @-> int @-> int @-> int @-> returning string_opt)

The return value `string_opt` is constructed by the memory content of the second
parameter `ptr char`, however the corresponding OCaml value of `ptr char` is
GC-able (because we no longer reference it) before coping it to the return
value. So the return value may be incorrect.

Indeed this is bug of CTypes that it does not keep the OCaml variable alive
through the C binding invocation.

We add a function which uses the OCaml value corresponding to the `ptr char`
after the C binding to keep it alive through the C binding invocation.

Signed-off-by: Yang Qian <yang.qian@citrix.com>