CP-28936: Assert that a host has enough pCPUs to run a VM
This blocks starting, resuming, and migration (pool_migrate as well as
migrate_send in- and cross-pool) if the target host does not have at least as
many pCPUs as the VM has vCPUs.

A new API error HOST_NOT_ENOUGH_PCPUS is introduced for this case.

The check can be overridden by setting the platform key `vcpu-unrestricted` to
`true` or `1`.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>