CA-295092: cancel resume task if process_header raises an exception
emu-manager waits until process_header is finished, and the resume
thread in xenopsd is waiting first for emu-manager to finish, and then
for the process_header threads.
If process_headers raises an exception the resume deadlocks.

The cancel callback on the task already has the right logic to send an
abort command to emu-manager, so call that as soon as we know we got an
error.
We need to do this before we attempt to Event.sync, because that would
deadlock as the other thread is xenopsd is not waiting for a reply from
this thread yet (it is waiting for emu-manager to finish first).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>