CA-293085: Stop xapi-nbd before eject
Usually xapi-nbd stops cooperatively, using xapi to unplug any VBDs
before it dies. During pool eject, the host is removed from the
pool and a reboot then issued. When xapi-nbd was still running at
this point it tried to unplug the VBDs and got stuck in a retry
loop because the pool refused to give it a session because it had
been ejected.

The solution is to stop xapi-nbd before we remove ourselves from
the pool, which is what this PR does.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>