CA-293678 SXM in partially upgraded pool with pre-7.3 hosts fails migration checks
assert_can_migrate_sender is added in 7.3 along with some code reoganization.
So if in any chance this is called from master to slave in RPC but slave is
before 7.3, the slave will raise exception and hance fail the entire code due
to this unknown method.

In assert_can_migrate_sender, it is mostly regarding pgpu check, which may not
be necessary in older release, so just ignore this unknown method.

Signed-off-by: YarsinCitrix <yarsin.he@citrix.com>