CP-29070: introduce VM.set_NVRAM_EFI_variables for varstored
We do not want to allow users to change the NVRAM while the VM is
running, but varstored must be able to do this on behalf of the VM
itself, so introduce a new API for it to use.

We can't use VM.set_NVRAM here, because there could be multiple daemons
updating runtime state in the NVRAM (e.g. UEFI variables and vTPM state).
We do not want to use `remove_from_NVRAM`/`add_to_NVRAM` because if HA is
active there could be a situation where the VM is left with no EFI
variables at all.

The mutex is to protect against concurrent changes to NVRAM from UEFI
and vTPM, we know we can only make changes to those from the local host.
All changes from the user are forbidden while the VM is running.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>