CA-290696: Block VM migration cancelation just before suspending VM
VM live migration has 2 phases:
1. live copy memory pages
2. stop the source domain and copy the dirty pages

Once the second phase starts, we have no chance to clean cancell the live
migration because the cancelation will shutdown the VM.

So we mark the VM as not cancellable just before suspending the VM because starts
from this moment the cancelation of the task will cause the VM shut down.
We have to achieve the purpose by a callback because `TASK` cannot seen by
lowlevel module. Xenopsd will notify xapi after setting the task to not
cancellable.

Signed-off-by: Yang Qian <yang.qian@citrix.com>