Move to dune and fix opam dependencies
We were depending on `rpc` in opam, but actually using `rpclib`.
This had the undesired effect of installing both Async and Lwt, when
a dependency only asked for one but not the other.
The dependencies should be more fine grained now.

The opam dependencies were created based on output from:
`dune external-lib-deps -p <pkg> @install @runtest`

Signed-off-by: Edwin Török <edvin.torok@citrix.com>