CP-29169: call varstore-rm on UEFI VM clone
On clone certain UEFI variables may need to be cleared.
Call `varstore-rm -c <uuid>` to perform this.

The actual variables to clear are defined in the `/etc/xapi.d/efi-clone` directory,
however the format of NVRAM is managed by `varstored`, hence XAPI can't
clear them itself and needs to invoke the helper executable.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>