CP-29420: finer grained locking for xenopsd metadata
Xenopsd has a global metadata lock for creating/deleting and updating
the VM cache table.
This means that if you have a VM start that triggers a xenopsd metadata
push then all events from other VMs get blocked until this metadata push
finishes. On a slave this can be especially slow because it needs to
wait for the RPC from master.

This change improves bootstorm on slaves.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>