Confirm that the interface exists before calling `ip`
The logs are full of errors from commands such as `ip addr show` complaining
that the given interface does not exist. A Script_error exception would be
raised in such cases, which is usually caught and thrown away higher up the
stack. Still, the alarming log message is left behind, which often confuses
people.

Instead of calling out to the `ip` command and hoping for the best, we now
first check whether the interface actually exists, and raise the new
Interface_does_not_exist error if not.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>