Fix unit test after introduction of cancellable field
Also make the test less strict: we do not want to update the testfiles
everytime we introduce a new field with a default value
(would defeat the purpose of having them in the first place).
If unmarshal+marshal produces an Rpc.t that is not identical to the
original then try unmarshaling once again and compare the OCaml values:
if they are equal then consider the test a success.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>