Improve update_getty
This used to call `killall -q -HUP mingetty agetty`, which always returned `1`,
because `mingetty` and `agetty` would never both be running. Replace the
process names with `-r .*getty`, which is happy if just one process is killed.

Also use the existing `call_script` (moved to the top of the file and extended
a little) for more fine-grained error logging.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>