CA-254835: make Xen 4.8+ featuresets compatible with Xen 4.7
The following commits for Xen 4.8 changed cpuid featuresets a bit:

https://xenbits.xen.org/gitweb/?p=xen.git;a=commitdiff;h=20e92c97f904aa460e2223c30dcad36c234496b6
https://xenbits.xen.org/gitweb/?p=xen.git;a=commitdiff;h=38b1c3a5affbe3451f366576aef9c7149d62de04

This prevents VM migration from 4.7 to 4.8+ versions of Xen. The new
featuresets are more architecturally correct, but are not compatible
with the old ones from Xapi POV.

Fake up those feature bits for Xapi. It's safe to do and will be replaced
by the new CPUID/MSR policy work in the future.

Signed-off-by: Sergey Dyasli <sergey.dyasli@citrix.com>