CP-28301, CP-28659, CP-28662: add firmware type and NVRAM parameter for xenopsd
BIOS boot: if absent or set to `Bios`
UEFI boot: if set to ``Uefi _`

The field needs to have a default value, otherwise xenopsd fails to
import existing VMs that do not have this field.

The NVRAM will be used initially for UEFI guests to store:
`EFI-variables-on-boot` and `EFI-variables-backend`.
This is stored inside a record to be fully type safe (the field in XAPI is a string map).
However it is meant to be extensible in the future, hence a string map.

The actual EFI variables are not sent here, instead `varstored` will
read them directly from XAPI: this avoids the value getting out of sync
between XAPI and xenopsd.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>