CA-297297 Make create_row idempotent
If a row already exists, and all of the fields passed in are the same,
we should do nothing and return successfully.

If a row already exists, and any of the fields are different to what was
passed in, we should fail.
 - For this I've created a new exception - Integrity_violation.

I've also updated the database tests to cover these cases

Signed-off-by: Richard Davies <richard.davies@citrix.com>