CA-299944: Proxy requests to updates
Rather than going over http over the network to fetch the RPMs and yum
metadata,
connect to the local xapi and have the requests forwarded over stunnel.

This behaviour is activated by calling the (undocumented)
Pool_update.attach
API with the new parameter 'use_localhost_proxy' set to true. The
absence of
the parameter is treated as 'false' and xapi will fall back to the old
behaviour.

Cherry-pick from c42a7fe5

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>
Signed-off-by: Min Li <min.li1@citrix.com>