CP-29857: defer setting device-model until first VM.start
We want to support starting a UEFI template in BIOS mode from XenCenter
by just setting the firmware mode to Bios.
To do that we have to defer setting the device-model field to as late as
possible:
BIOS would have to default to qemu-upstream-compat,
and UEFI would have to default to qemu-upstream-uefi.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>