CP-29936: override control/feature-suspend if data/cant_suspend_reason is set
The guest might be able to suspend, but QEMU may not support it,
for example if the guest has not unplugged its NVME disks yet.
Remove the "feature-suspend" from guest metrics and replace it with
the reason error from QEMU when this happens, this will cause the
allowed operations on the VM to be updated accordingly.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>