CP-29967: varstored deprivileging
Spawn varstored in a chroot and run it as non-root, like we already do
with QEMU.
This requires carefully managing paths, because the ones we pass to
varstored need to be relative to the chroot, but xenopsd needs to use
absolute paths. The Chroot module tries to prevent accidental misuse of
relative and absolute paths.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>