CA-303253: XenMotion breaks VIF connectivity when using DVSC
When migrating a VM whose network managed by DVSC, the network connection of the
VM will broken after the migration.

This is caused by we name the target VM of the migration with a temporary UUID
and we store the temporary UUID in OVS DB which will be used by DVSC. However
when finishing the migration, we will rename the UUID of the target VM to the
original VM UUID and will lead DVSC unable to find the VM by the temporary UUID
in OVS DB.

To fix this issue, we introduce a new key /VM/<VM-UUID>/final-uuid in xenstore.
When creating the migration target VM, we will set the original UUID as the
value of this key. The script of setting up network for the VM will read this
key and process it properly. When the migration finished, we will delete the key
in xenstore.

Signed-off-by: Yang Qian <yang.qian@citrix.com>