Use dune and define profile "gprof" for profiling
Switch to to Dune and define a profile for profiling with gprof while
keeping jbuild files for now. Since Dune has stricter warning flags in
its dev profile, we have to use the release profile to avoid braking the
build until this is cleaned up. Switching to a different profile is easy
from the command line:

  make PROFILE=dev

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>