CP-29827: drop some trad-compat options (#579)
* CP-29827: drop some trad-compat options

In UEFI boot mode we get a chance to start fresh with a new version of
QEMU device-model, so move all the code that is needed for backwards compatibility
into the qemu-upstream-compat profile.
Always add romfile=

More uniform parameter building, instead of special cased string
manipulation.

pass --device to vgpu in UEFI mode: demu would default to 0x0b if it is not passed.
We can't change that due to backwards compatibility reasons for BIOS,
but we can change it for UEFI.
In UEFI mode the vGPU device numbers will start at 8.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>