CA-308072: do not swallow errors from creating the http server
For example I had an EADDRINUSE error while developing the toolstack
restart code, and only way to find it was via strace, there was nothing
in the logs.

Change this so instead we wait until the first of these events:
 * the socket is present and connectable
 * creating the server raises an exception

In the former case we exit the function cleanly, in the latter case we
cancel the loop that waits for the socket to become connectable
(it never will).
This is done by using Lwt.pick, which waits until one Lwt promise is
resolved and attempts to cancel all others.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>