CA-306395 Convert sendfile to direct_copy
In the general case it is better to use read() and write() with O_DIRECT
for large data transfers than to unconditionally send everything via
sendfile(), particularly when the output is not a pipe or a socket.

This converts the previous sendfile wrapper into a more general
direct_copy wrapper. Later, if we decide we need to get more performance
between certain types of filedescriptors, we can make use of the
stub_init() function which currently allocates the buffer record an
engine selector in the direct_copy_handle based on information obtained
from fstat() on the two filedescriptors.

Incorporates changes from Rob Hoes <rob.hoes@citrix.com>:
    Remove unused reference to sendfile binding
    Add init and cleanup stubs for sendfile replacement

Signed-off-by: Tim Smith <tim.smith@citrix.com>
Signed-off-by: Rob Hoes <rob.hoes@citrix.com>