CP-29962: Ignore monitor_config_file for GVT-g VGPU types
This information is no longer required for starting GVT-g VMs, so xapi no
longer needs to read it from the whitelist file and pass it to xenopsd. Note
that, for backwards compatibility, we need to keep the `monitor_config_file`
key in the `Xenops_interface`, but xapi sets it to `None` from now on.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>