CA-309024 improve SIGTERM signal handling
The signal handler for SIGTERM raises Killed. This exception can escape
in the presence of threads and may be ignored because it is not visible
outside. This commit adds global reference killed which is set by the
signal handler and checked in the main loop such that we are guaranteed
to break it in the presence of SIGTERM and perform all cleanup actions.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>