CA-309427: interpret the -serial option rather than looking for hvm_serial
The 'hvm_serial' option can be set on either the 'platform' or
'other-config' keys. Only the former causes something to be written into
xenstore.
Also, what we're trying to do is avoid the network unshare in the case that
a serial device is re-directed to the network. Hence the wrapper should
be parsing any '-serial' argument looking for a network re-direct. In
practice the only one we care about is 'tcp' so it's sufficient to look for
that prefix. If other types of re-direction become commonplace in future
then the test can be extended.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>