CA-309685 fix fd leak in QMP connection handling
In the case of QMP negotiation failure, remove() is called but it does
not know the connection that it needs to close. This commit makes sure
remove() does know it by adding the connection to the lookup table
before calling Qmp_protocol.negotiate() (which might fail and trigger
the cleanup case).

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>